comment form list & files

+++ lib/screens/forum/widgets/post.view.dart
@@ -1,7 +1,10 @@
 import 'package:flutter/material.dart';
+import 'package:nalia_app/models/api.comment.model.dart';
 import 'package:nalia_app/models/api.controller.dart';
 import 'package:nalia_app/models/api.post.model.dart';
-import 'package:nalia_app/screens/forum/widgets/post.view.files.dart';
+import 'package:nalia_app/screens/forum/widgets/comment.form.dart';
+import 'package:nalia_app/screens/forum/widgets/comment.list.dart';
+import 'package:nalia_app/screens/forum/widgets/files.view.dart';
 import 'package:nalia_app/services/defines.dart';
 import 'package:nalia_app/services/global.dart';
 
@@ -24,6 +27,7 @@ class _PostViewState extends State<PostView> {
   Widget build(BuildContext context) {
     Forum forum = widget.forum;
     ApiPost post = widget.forum.posts[widget.i];
+
     return Container(
       width: double.infinity,
       margin: EdgeInsets.all(sm),
@@ -35,33 +39,41 @@ class _PostViewState extends State<PostView> {
           Text('Post No. ${post.id}, Author: ${post.authorName}'),
           Text('title: ${post.postTitle}'),
           Text('content: ${post.postContent}'),
-          PostViewFiles(post: post),
+          PostViewFiles(postOrComment: post),
           Row(
             children: [
-              RaisedButton(
-                child: Text('Edit'),
-                onPressed: () {
-                  if (post.isNotMine) return app.alert('not your post');
-                  forum.editPost(post);
-                },
-              ),
-              RaisedButton(
-                child: Text('Delete'),
-                onPressed: () async {
-                  try {
-                    final re = await app.confirm(
-                        'Delete', 'Do you want to delete the post?');
-                    if (re == false) return;
-                    final deletedId = await api.deletePost(post, forum);
-                    print('deletedId: $deletedId');
-                    forum.render();
-                  } catch (e) {
-                    app.error(e);
-                  }
-                },
-              ),
+              if (post.isMine)
+                RaisedButton(
+                  child: Text('Edit'),
+                  onPressed: () {
+                    if (post.isNotMine) return app.alert('not your post');
+                    forum.editPost(post);
+                  },
+                ),
+              if (post.isMine)
+                RaisedButton(
+                  child: Text('Delete'),
+                  onPressed: () async {
+                    try {
+                      final re = await app.confirm(
+                          'Delete', 'Do you want to delete the post?');
+                      if (re == false) return;
+                      final deletedId = await api.deletePost(post, forum);
+                      print('deletedId: $deletedId');
+                      forum.render();
+                    } catch (e) {
+                      app.error(e);
+                    }
+                  },
+                ),
             ],
-          )
+          ),
+          CommentForm(
+            post: post,
+            forum: forum,
+            comment: ApiComment(),
+          ),
+          CommentList(post: post, forum: forum),
         ],
       ),
     );
